feat: reducer 폴더 생성 및 ballReducer 분리
- min27604/baseball/#15
