+++ FE/react-baseball/src/App.jsx
@@ -1,7 +1,7 @@
+import { useReducer, useState, createContext } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { createGlobalStyle, ThemeProvider } from 'styled-components';
-import React, { useReducer, useState, createContext } from 'react';
-import styled from 'styled-components';
+import styled, { createGlobalStyle, ThemeProvider } from 'styled-components';
+import ballReducer from './components/reducer/ballReducer';
 import {
   boardHistory,
   BoardHistoryContext,
@@ -11,20 +11,8 @@ import PlayScreen from './routes/PlayScreen';
 import StartScreen from './routes/StartScreen';
 import theme from './theme';
 
-const ballReducer = (ballCnt, action) => {
-  switch (action.type) {
-    case 'S':
-      return { ...ballCnt, S: ballCnt.S + 1 };
-    case 'B':
-      return { ...ballCnt, B: ballCnt.B + 1 };
-    case 'O':
-      return { ...ballCnt, O: ballCnt.O + 1 };
-  }
-};
-
 function App() {
   const [ballCnt, dispatch] = useReducer(ballReducer, boardHistory);
-  const { Provider } = BoardHistoryContext;
   const [playable, setPlayable] = useState(true);
 
   return (
@@ -37,11 +25,11 @@ function App() {
               <Route path='/' exact>
                 <StartScreen />
               </Route>
-              <Provider value={{ ballCnt, dispatch }}>
+              <BoardHistoryContext.Provider value={{ ballCnt, dispatch }}>
                 <Route path='/play-screen'>
                   <PlayScreen />
                 </Route>
-              </Provider>
+              </BoardHistoryContext.Provider>
             </AppDiv>
           </Switch>
         </BrowserRouter>
@@ -56,7 +44,6 @@ const AppDiv = styled.div`
 `;
 const GlobalStyle = createGlobalStyle`
 * {
-
   padding:0; 
   margin:0;
 }
@@ -72,5 +59,4 @@ ol, ul {
 `;
 
 export const PlayableContext = createContext();
-
 export default App;
